City: Introduce click actions
Each cell may have a click action associated with it.

Eg the build city button has the `BuildCityClickAction`.

The `SelectHexClickAction` is the default click action.
This simply highlights the clicked map cell.

Now, a user can select a cell, and then place a city by clicking the build
city button.